apps/cli: Split telegram-client into per-transport modules with a respond router
The mobile and Telegram message shapes were both living inside telegram-client.ts
alongside the dispatch logic, which made the file the de-facto home for the
entire wire protocol. Move the studio-mobile shape into its own
studio-mobile-client.ts, the Telegram-only code stays in telegram-client.ts,
and introduce respond-router.ts as the dispatcher (fork + retry loop). Shared
HTTP primitives (errors, URL helpers, backoff) move to remote-http.ts; the
Telegram-flavored error class names become Remote*Error to reflect that they're
transport-layer concerns shared by both clients.
